Move Espoo client to profile-activated config
